[automation] Extract port opening program buying into a new module
